improvement for grpcx (#2510)
