+++ example/trace/grpc_with_db/client/main.go
@@ -1,12 +1,12 @@
 package main
 
 import (
+	"github.com/gogf/gf/contrib/rpc/grpcx/v2"
 	"github.com/gogf/gf/contrib/trace/jaeger/v2"
+	"github.com/gogf/gf/example/trace/grpc_with_db/protocol/user"
 	"github.com/gogf/gf/v2/frame/g"
 	"github.com/gogf/gf/v2/net/gtrace"
 	"github.com/gogf/gf/v2/os/gctx"
-
-	"github.com/gogf/gf/example/trace/grpc_with_db/protobuf/user"
 )
 
 const (
@@ -29,16 +29,13 @@ func StartRequests() {
 	ctx, span := gtrace.NewSpan(gctx.New(), "StartRequests")
 	defer span.End()
 
-	var client, err = user.NewClient()
-	if err != nil {
-		g.Log().Fatalf(ctx, `%+v`, err)
-	}
+	client := user.NewUserClient(grpcx.Client.MustNewGrpcClientConn("demo"))
 
 	// Baggage.
 	ctx = gtrace.SetBaggageValue(ctx, "uid", 100)
 
 	// Insert.
-	insertRes, err := client.User().Insert(ctx, &user.InsertReq{
+	insertRes, err := client.Insert(ctx, &user.InsertReq{
 		Name: "john",
 	})
 	if err != nil {
@@ -47,7 +44,7 @@ func StartRequests() {
 	g.Log().Info(ctx, "insert id:", insertRes.Id)
 
 	// Query.
-	queryRes, err := client.User().Query(ctx, &user.QueryReq{
+	queryRes, err := client.Query(ctx, &user.QueryReq{
 		Id: insertRes.Id,
 	})
 	if err != nil {
@@ -57,7 +54,7 @@ func StartRequests() {
 	g.Log().Info(ctx, "query result:", queryRes)
 
 	// Delete.
-	_, err = client.User().Delete(ctx, &user.DeleteReq{
+	_, err = client.Delete(ctx, &user.DeleteReq{
 		Id: insertRes.Id,
 	})
 	if err != nil {
@@ -67,7 +64,7 @@ func StartRequests() {
 	g.Log().Info(ctx, "delete id:", insertRes.Id)
 
 	// Delete with error.
-	_, err = client.User().Delete(ctx, &user.DeleteReq{
+	_, err = client.Delete(ctx, &user.DeleteReq{
 		Id: -1,
 	})
 	if err != nil {
