+++ example/trace/grpc_with_db/protobuf/user/user.proto
