+++ example/trace/grpc_with_db/server/config.yaml
@@ -1,8 +1,6 @@
 
-registry:
-  endpoints: ["127.0.0.1:2379"]
 
-grpcserver:
+grpc:
   name:            "demo"
   logStdout:        true
   errorLogEnabled:  true
